requirements.txt
Django==6.0.1
djangorestframework==3.15.1
django-cors-headers==4.4.0
PyMySQL==1.1.0